feat(cmd/scaffold): add module, wasm, query, message, packat cmds
