+++ starport/cmd/cmd.go
@@ -40,13 +40,10 @@ func New() *cobra.Command {
 	c.AddCommand(NewChain())
 	c.AddCommand(NewDocs())
 	c.AddCommand(NewType())
-	c.AddCommand(NewModule())
+	c.AddCommand(NewScaffoldModule())
 	c.AddCommand(NewRelayer())
 	c.AddCommand(NewVersion())
 	c.AddCommand(NewNetwork())
-	c.AddCommand(NewIBCPacket())
-	c.AddCommand(NewMessage())
-	c.AddCommand(NewQuery())
 	c.AddCommand(NewTools())
 	c.AddCommand(deprecated()...)
 	return c
