+++ starport/cmd/scaffold_module.go
@@ -22,16 +22,11 @@ const (
 
 var ibcRouterPlaceholderInstruction = fmt.Sprintf(`
 💬 To enable scaffolding of IBC modules, remove these lines from app/app.go:
-
 %s
-
 💬 Then, find the following line:
-
 %s
-
 💬 Finally, add this block of code below:
 %s
-
 `,
 	infoColor(`ibcRouter := porttypes.NewRouter()
 ibcRouter.AddRoute(ibctransfertypes.ModuleName, transferModule)
@@ -43,15 +38,15 @@ ibcRouter.AddRoute(ibctransfertypes.ModuleName, transferModule)
 app.IBCKeeper.SetRouter(ibcRouter)`),
 )
 
-// NewModuleCreate creates a new module create command to scaffold an
+// NewScaffoldModule  creates a new module create command to scaffold an
 // sdk module.
-func NewModuleCreate() *cobra.Command {
+func NewScaffoldModule() *cobra.Command {
 	c := &cobra.Command{
-		Use:   "create [name]",
+		Use:   "module [name]",
 		Short: "Scaffold a Cosmos SDK module",
 		Long:  "Scaffold a new Cosmos SDK module in the `x` directory",
 		Args:  cobra.MinimumNArgs(1),
-		RunE:  createModuleHandler,
+		RunE:  scaffoldModuleHandler,
 	}
 	c.Flags().Bool(flagIBC, false, "scaffold an IBC module")
 	c.Flags().String(flagIBCOrdering, "none", "channel ordering of the IBC module [none|ordered|unordered]")
@@ -59,7 +54,7 @@ func NewModuleCreate() *cobra.Command {
 	return c
 }
 
-func createModuleHandler(cmd *cobra.Command, args []string) error {
+func scaffoldModuleHandler(cmd *cobra.Command, args []string) error {
 	s := clispinner.New().SetText("Scaffolding...")
 	defer s.Stop()
