+++ starport/cmd/scaffold_query.go
@@ -10,27 +10,27 @@ import (
 )
 
 const (
-	flagResponse    string = "response"
-	flagDescription string = "desc"
+	flagPaginated = "paginated"
 )
 
-// NewType command creates a new type command to scaffold messages
-func NewMessage() *cobra.Command {
+// NewScaffoldQuery command creates a new type command to scaffold queries
+func NewScaffoldQuery() *cobra.Command {
 	c := &cobra.Command{
-		Use:   "message [name] [field1] [field2] ...",
-		Short: "Scaffold a Cosmos SDK message",
+		Use:   "query [name] [request_field1] [request_field2] ...",
+		Short: "Scaffold a Cosmos SDK query",
 		Args:  cobra.MinimumNArgs(1),
-		RunE:  messageHandler,
+		RunE:  queryHandler,
 	}
 	c.Flags().StringVarP(&appPath, "path", "p", "", "path of the app")
-	c.Flags().String(flagModule, "", "Module to add the message into. Default: app's main module")
+	c.Flags().String(flagModule, "", "Module to add the query into. Default: app's main module")
 	c.Flags().StringSliceP(flagResponse, "r", []string{}, "Response fields")
 	c.Flags().StringP(flagDescription, "d", "", "Description of the command")
+	c.Flags().Bool(flagPaginated, false, "Define if the request can be paginated")
 
 	return c
 }
 
-func messageHandler(cmd *cobra.Command, args []string) error {
+func queryHandler(cmd *cobra.Command, args []string) error {
 	s := clispinner.New().SetText("Scaffolding...")
 	defer s.Stop()
 
@@ -40,7 +40,7 @@ func messageHandler(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	// Get response fields
+	// Get request fields
 	resFields, err := cmd.Flags().GetStringSlice(flagResponse)
 	if err != nil {
 		return err
@@ -53,14 +53,19 @@ func messageHandler(cmd *cobra.Command, args []string) error {
 	}
 	if desc == "" {
 		// Use a default description
-		desc = fmt.Sprintf("Broadcast message %s", args[0])
+		desc = fmt.Sprintf("Query %s", args[0])
+	}
+
+	paginated, err := cmd.Flags().GetBool(flagPaginated)
+	if err != nil {
+		return err
 	}
 
 	sc, err := scaffolder.New(appPath)
 	if err != nil {
 		return err
 	}
-	sm, err := sc.AddMessage(placeholder.New(), module, args[0], desc, args[1:], resFields)
+	sm, err := sc.AddQuery(placeholder.New(), module, args[0], desc, args[1:], resFields, paginated)
 	if err != nil {
 		return err
 	}
@@ -68,6 +73,6 @@ func messageHandler(cmd *cobra.Command, args []string) error {
 	s.Stop()
 
 	fmt.Println(sourceModificationToString(sm))
-	fmt.Printf("\n🎉 Created a message `%[1]v`.\n\n", args[0])
+	fmt.Printf("\n🎉 Created a query `%[1]v`.\n\n", args[0])
 	return nil
 }
